enterprise-agent/pyproject.toml
[project]
name = "enterprise-agent"
version = "0.1.0"
description = "Production-grade multi-agent AI system"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "anthropic>=0.40.0",
    "tenacity>=8.2.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "prometheus-client>=0.20.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "locust>=2.24.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
